feat: rework style props and transform functions
