+++ packages/docs/pages/borders.mdx
@@ -3,18 +3,27 @@ import ThemeParser from "../components/ThemeParser";
 
 # Borders
 
-The border theme object is around the following scales:
 The border theme object uses the following scales that are based on the [Styled System Theme Specification](https://styled-system.com/theme-specification):
 
-`borders`
-
-`borderWidths`
-
-`radii` (Radius)
+* `borders`
+* `borderStyles`
+* `borderWidths`
+* `radii` (`border-radius`)
 
 ## Border
 
+Defaults to an empty object.
+
 <ThemeParser theme="borders" />
+
+## Border Style
+
+Defaults to an empty object.
+
+<ThemeParser theme="borderStyles" />
+
+## Border Width
+
 <ThemeParser theme="borderWidths" />
 
 ```jsx
