+++ packages/docs/pages/typography.mdx
@@ -5,7 +5,7 @@ import ThemeParser from "../components/ThemeParser";
 
 To manage Typography options, the theme object supports the following keys:
 
-- `fonts` (font family)
+- `fonts` (`font-family`)
 - `fontSizes`
 - `fontWeights`
 - `lineHeights`
