--- packages/react/src/shared/styled-system/background.js
@@ -1,18 +1,0 @@
-import { system } from '@styled-system/core';
-
-const config = {
-  background: true,
-  backgroundImage: true,
-  backgroundSize: true,
-  backgroundPosition: true,
-  backgroundRepeat: true,
-};
-
-config.bgImage = config.backgroundImage;
-config.bgSize = config.backgroundSize;
-config.bgPosition = config.backgroundPosition;
-config.bgRepeat = config.backgroundRepeat;
-
-export const background = system(config);
-
-export default background;
