--- packages/react/src/shared/styled-system/color.js
@@ -1,19 +1,0 @@
-import { system } from '@styled-system/core';
-
-const config = {
-  color: {
-    property: 'color',
-    scale: 'colors',
-  },
-  backgroundColor: {
-    property: 'backgroundColor',
-    scale: 'colors',
-  },
-  opacity: true,
-}
-
-config.bg = config.backgroundColor;
-
-export const color = system(config);
-
-export default color;
