--- packages/react/src/shared/styled-system/config.js
@@ -1,99 +1,0 @@
-const config = {
-  // Animation
-  animation: true,
-  animationDelay: true,
-  animationDirection: true,
-  animationDuration: true,
-  animationFillMode: true,
-  animationIterationCount: true,
-  animationName: true,
-  animationPlayState: true,
-  animationTimingFunction: true,
-
-  // Background
-  backgroundAttachment: true,
-  backgroundClip: true,
-
-  // Color
-  colorScheme: true,
-
-  // Containment
-  contain: true,
-  containIntrinsicSize: true,
-  contentVisibility: true,
-
-  // Interactivity
-  appearance: true,
-  caretColor: true,
-  cursor: true,
-  outline: true,
-  outlineColor: true,
-  outlineOffset: true,
-  pointerEvents: true,
-  resize: true,
-  userSelect: true,
-
-  // Layout
-  aspectRatio: true,
-  boxSizing: true,
-  float: true,
-  objectFit: true,
-  objectPosition: true,
-  visibility: true,
-
-  // List
-  listStyleImage: true,
-  listStylePosition: true,
-  listStyleType: true,
-
-  // Performance
-  willChange: true,
-
-  // SVG
-  fill: {
-    property: 'fill',
-    scale: 'colors',
-  },
-  stroke: {
-    property: 'stroke',
-    scale: 'colors',
-  },
-
-  // Text decoration
-  textDecoration: true,
-  textDecorationColor: true,
-  textDecorationLine: true,
-  textDecorationStyle: true,
-  textDecorationThickness: true,
-  textUnderlineOffset: true,
-
-  // Transform
-  backfaceVisibility: true,
-  perspective: true,
-  perspectiveOrigin: true,
-  transform: true,
-  transformBox: true,
-  transformOrigin: true,
-  transformStyle: true,
-
-  // Transition
-  transition: true,
-  transitionDelay: true,
-  transitionDuration: true,
-  transitionProperty: true,
-  transitionTimingFunction: true,
-
-  // Typography
-  lineBreak: true,
-  overflowWrap: true,
-  textIndent: true,
-  textJustify: true,
-  textTransform: true,
-  textEmphasis: true,
-  textOverflow: true,
-  whiteSpace: true,
-  wordBreak: true,
-  wordSpacing: true,
-};
-
-export default config;
