+++ packages/react/src/shared/styled-system/config/animation.js
@@ -1,0 +1,17 @@
+import { system } from '@styled-system/core';
+
+const config = {
+  animation: true,
+  animationDelay: true,
+  animationDirection: true,
+  animationDuration: true,
+  animationFillMode: true,
+  animationIterationCount: true,
+  animationName: true,
+  animationPlayState: true,
+  animationTimingFunction: true,
+};
+
+const animation = system(config);
+
+export default animation;
