+++ packages/react/src/shared/styled-system/config/containment.js
@@ -1,0 +1,11 @@
+import { system } from '@styled-system/core';
+
+const config = {
+  contain: true,
+  containIntrinsicSize: true,
+  contentVisibility: true,
+};
+
+const containment = system(config);
+
+export default containment;
