+++ packages/react/src/shared/styled-system/config/interactivity.js
@@ -1,0 +1,17 @@
+import { system } from '@styled-system/core';
+
+const config = {
+  appearance: true,
+  caretColor: true,
+  cursor: true,
+  outline: true,
+  outlineColor: true,
+  outlineOffset: true,
+  pointerEvents: true,
+  resize: true,
+  userSelect: true,
+};
+
+const interactivity = system(config);
+
+export default interactivity;
