+++ packages/react/src/shared/styled-system/config/list-style.js
@@ -1,0 +1,11 @@
+import { system } from '@styled-system/core';
+
+const config = {
+  listStyleImage: true,
+  listStylePosition: true,
+  listStyleType: true,
+};
+
+const listStyle = system(config);
+
+export default listStyle;
