+++ packages/react/src/shared/styled-system/config/position.js
@@ -1,4 +1,5 @@
 import { system } from '@styled-system/core';
+import * as transform from '../utils/transform-functions';
 
 const config = {
   position: true,
@@ -9,21 +10,25 @@ const config = {
   top: {
     property: 'top',
     scale: 'space',
+    transform: transform.margin,
   },
   right: {
     property: 'right',
     scale: 'space',
+    transform: transform.margin,
   },
   bottom: {
     property: 'bottom',
     scale: 'space',
+    transform: transform.margin,
   },
   left: {
     property: 'left',
     scale: 'space',
+    transform: transform.margin,
   },
 };
 
-export const position = system(config);
+const position = system(config);
 
 export default position;
