+++ packages/react/src/shared/styled-system/config/text-decoration.js
@@ -1,0 +1,18 @@
+import { system } from '@styled-system/core';
+
+const config = {
+  textDecoration: true,
+  textDecorationColor: true,
+  textDecorationLine: true,
+  textDecorationStyle: true,
+  textDecorationThickness: true,
+  textShadow: {
+    property: 'textShadow',
+    scale: 'shadows',
+  },
+  textUnderlineOffset: true,
+};
+
+const textDecoration = system(config);
+
+export default textDecoration;
