+++ packages/react/src/shared/styled-system/config/transform.js
@@ -1,0 +1,15 @@
+import { system } from '@styled-system/core';
+
+const config = {
+  backfaceVisibility: true,
+  perspective: true,
+  perspectiveOrigin: true,
+  transform: true,
+  transformBox: true,
+  transformOrigin: true,
+  transformStyle: true,
+};
+
+const transform = system(config);
+
+export default transform;
