+++ packages/theme/src/foundations/borderWidths.js
@@ -1,7 +1,3 @@
-const borderWidths = {
-  none: 0,
-  1: '1px',
-  2: '2px',
-};
+const borderWidths = {};
 
 export default borderWidths;
