+++ packages/theme/src/foundations/radii.js
@@ -1,9 +1,3 @@
-const radii = {
-  circle: '50%',
-  none: 0,
-  sm: '.1875rem',
-  md: '.375rem',
-  lg: '.75rem',
-};
+const radii = {};
 
 export default radii;
